+++ src/ska_sdp_func_python/calibration/__init__.py
@@ -3,6 +3,7 @@
 Calibration of observations, using single Jones matricies
 or chains of Jones matrices.
 """
+from .beamformer_utils import *
 from .chain_calibration import *
 from .dp3_calibration import *
 from .jones import *